Add Boilerplate for chroot testing
Add a constant CHROOT_PREFIX to stem/socket.py which
contains the prefix to stripped out. ControlMessage.__init__()
checks if the CHROOT_PREFIX is set. If set, then
ControlMEssage.strip() is called which strips out the prefix.
Set CHROOT_PREFIX=get_test_dir() in Runner.start() and remove it
in Runner.stop()